Make separate file for distribution logic
