+++ x/incentives/keeper/gauge.go
@@ -43,31 +43,8 @@ func (k Keeper) getCoinsFromGauges(gauges []types.Gauge) sdk.Coins {
 	return coins
 }
 
-func (k Keeper) getDistributedCoinsFromGauges(gauges []types.Gauge) sdk.Coins {
-	coins := sdk.Coins{}
-	for _, gauge := range gauges {
-		coins = coins.Add(gauge.DistributedCoins...)
-	}
-	return coins
-}
-
-func (k Keeper) getToDistributeCoinsFromGauges(gauges []types.Gauge) sdk.Coins {
-	// TODO: Consider optimizing this in the future to only require one iteration over all gauges.
-	coins := k.getCoinsFromGauges(gauges)
-	distributed := k.getDistributedCoinsFromGauges(gauges)
-	return coins.Sub(distributed)
-}
-
 func (k Keeper) getCoinsFromIterator(ctx sdk.Context, iterator db.Iterator) sdk.Coins {
 	return k.getCoinsFromGauges(k.getGaugesFromIterator(ctx, iterator))
-}
-
-func (k Keeper) getToDistributeCoinsFromIterator(ctx sdk.Context, iterator db.Iterator) sdk.Coins {
-	return k.getToDistributeCoinsFromGauges(k.getGaugesFromIterator(ctx, iterator))
-}
-
-func (k Keeper) getDistributedCoinsFromIterator(ctx sdk.Context, iterator db.Iterator) sdk.Coins {
-	return k.getDistributedCoinsFromGauges(k.getGaugesFromIterator(ctx, iterator))
 }
 
 // setGauge set the gauge inside store
@@ -180,287 +157,6 @@ func (k Keeper) AddToGaugeRewards(ctx sdk.Context, owner sdk.AccAddress, coins s
 	}
 	k.hooks.AfterAddToGauge(ctx, gauge.Id)
 	return nil
-}
-
-// BeginDistribution is a utility to begin distribution for a specific gauge
-func (k Keeper) BeginDistribution(ctx sdk.Context, gauge types.Gauge) error {
-	// validation for current time and distribution start time
-	if ctx.BlockTime().Before(gauge.StartTime) {
-		return fmt.Errorf("gauge is not able to start distribution yet: %s >= %s", ctx.BlockTime().String(), gauge.StartTime.String())
-	}
-
-	timeKey := getTimeKey(gauge.StartTime)
-	if err := k.deleteGaugeRefByKey(ctx, combineKeys(types.KeyPrefixUpcomingGauges, timeKey), gauge.Id); err != nil {
-		return err
-	}
-	if err := k.addGaugeRefByKey(ctx, combineKeys(types.KeyPrefixActiveGauges, timeKey), gauge.Id); err != nil {
-		return err
-	}
-	return nil
-}
-
-// FinishDistribution is a utility to finish distribution for a specific gauge
-func (k Keeper) FinishDistribution(ctx sdk.Context, gauge types.Gauge) error {
-	timeKey := getTimeKey(gauge.StartTime)
-	if err := k.deleteGaugeRefByKey(ctx, combineKeys(types.KeyPrefixActiveGauges, timeKey), gauge.Id); err != nil {
-		return err
-	}
-	if err := k.addGaugeRefByKey(ctx, combineKeys(types.KeyPrefixFinishedGauges, timeKey), gauge.Id); err != nil {
-		return err
-	}
-	if err := k.deleteGaugeIDForDenom(ctx, gauge.Id, gauge.DistributeTo.Denom); err != nil {
-		return err
-	}
-	k.hooks.AfterFinishDistribution(ctx, gauge.Id)
-	return nil
-}
-
-// GetLocksToDistribution get locks that are associated to a condition
-func (k Keeper) GetLocksToDistribution(ctx sdk.Context, distrTo lockuptypes.QueryCondition) []lockuptypes.PeriodLock {
-	switch distrTo.LockQueryType {
-	case lockuptypes.ByDuration:
-		return k.lk.GetLocksLongerThanDurationDenom(ctx, distrTo.Denom, distrTo.Duration)
-	case lockuptypes.ByTime:
-		return k.lk.GetLocksPastTimeDenom(ctx, distrTo.Denom, distrTo.Timestamp)
-	default:
-	}
-	return []lockuptypes.PeriodLock{}
-}
-
-// getLocksToDistributionWithMaxDuration get locks that are associated to a condition
-// and if its by duration, then use the min Duration
-func (k Keeper) getLocksToDistributionWithMaxDuration(ctx sdk.Context, distrTo lockuptypes.QueryCondition, minDuration time.Duration) []lockuptypes.PeriodLock {
-	switch distrTo.LockQueryType {
-	case lockuptypes.ByDuration:
-		if distrTo.Duration > minDuration {
-			return k.lk.GetLocksLongerThanDurationDenom(ctx, distrTo.Denom, minDuration)
-		}
-		return k.lk.GetLocksLongerThanDurationDenom(ctx, distrTo.Denom, distrTo.Duration)
-	case lockuptypes.ByTime:
-		panic("Gauge by time is present!?!? Should have been blocked in ValidateBasic")
-	default:
-	}
-	return []lockuptypes.PeriodLock{}
-}
-
-// FilteredLocksDistributionEst estimate distribution amount coins from gauge for fitting conditions
-// Expectation: gauge is a valid gauge
-// filteredLocks are all locks that are valid for gauge
-// It also applies an update for the gauge, handling the sending of the rewards.
-// (Note this update is in-memory, it does not change state.)
-func (k Keeper) FilteredLocksDistributionEst(ctx sdk.Context, gauge types.Gauge, filteredLocks []lockuptypes.PeriodLock) (types.Gauge, sdk.Coins, error) {
-	TotalAmtLocked := k.lk.GetPeriodLocksAccumulation(ctx, gauge.DistributeTo)
-	if TotalAmtLocked.IsZero() {
-		return types.Gauge{}, nil, nil
-	}
-
-	remainCoins := gauge.Coins.Sub(gauge.DistributedCoins)
-	// Remaining epochs is the number of remaining epochs that the gauge will pay out its rewards
-	// For a perpetual gauge, it will pay out everything in the next epoch, and we don't make
-	// an assumption for what rate it will get refilled at.
-	remainEpochs := uint64(1)
-	if !gauge.IsPerpetual {
-		remainEpochs = gauge.NumEpochsPaidOver - gauge.FilledEpochs
-	}
-	// TODO: Should this return err
-	if remainEpochs == 0 {
-		return gauge, sdk.Coins{}, nil
-	}
-
-	remainCoinsPerEpoch := sdk.Coins{}
-	for _, coin := range remainCoins {
-		// distribution amount per epoch = gauge_size / (remain_epochs)
-		amt := coin.Amount.QuoRaw(int64(remainEpochs))
-		remainCoinsPerEpoch = remainCoinsPerEpoch.Add(sdk.NewCoin(coin.Denom, amt))
-	}
-
-	// Now we compute the filtered coins
-	filteredDistrCoins := sdk.Coins{}
-	if len(filteredLocks) == 0 {
-		// If were doing no filtering, we want to calculate the total amount to distributed in
-		// the next epoch.
-		// distribution in next epoch = gauge_size  / (remain_epochs)
-		filteredDistrCoins = remainCoinsPerEpoch
-	}
-	for _, lock := range filteredLocks {
-		denomLockAmt := lock.Coins.AmountOf(gauge.DistributeTo.Denom)
-
-		for _, coin := range remainCoinsPerEpoch {
-			// distribution amount = gauge_size * denom_lock_amount / (total_denom_lock_amount * remain_epochs)
-			// distribution amount = gauge_size_per_epoch * denom_lock_amount / total_denom_lock_amount
-			amt := coin.Amount.Mul(denomLockAmt).Quo(TotalAmtLocked)
-			filteredDistrCoins = filteredDistrCoins.Add(sdk.NewCoin(coin.Denom, amt))
-		}
-	}
-
-	// increase filled epochs after distribution
-	gauge.FilledEpochs += 1
-	gauge.DistributedCoins = gauge.DistributedCoins.Add(remainCoinsPerEpoch...)
-
-	return gauge, filteredDistrCoins, nil
-}
-
-// distributionInfo stores all of the information for pent up sends for rewards distributions.
-// This enables us to lower the number of events and calls to back
-type distributionInfo struct {
-	nextID            int
-	lockOwnerAddrToID map[string]int
-	idToBech32Addr    []string
-	idToDecodedAddr   []sdk.AccAddress
-	idToDistrCoins    []sdk.Coins
-}
-
-func newDistributionInfo() distributionInfo {
-	return distributionInfo{
-		nextID:            0,
-		lockOwnerAddrToID: make(map[string]int),
-		idToBech32Addr:    []string{},
-		idToDecodedAddr:   []sdk.AccAddress{},
-		idToDistrCoins:    []sdk.Coins{},
-	}
-}
-
-func (d *distributionInfo) addLockRewards(lock lockuptypes.PeriodLock, rewards sdk.Coins) error {
-	if id, ok := d.lockOwnerAddrToID[lock.Owner]; ok {
-		oldDistrCoins := d.idToDistrCoins[id]
-		d.idToDistrCoins[id] = rewards.Add(oldDistrCoins...)
-	} else {
-		id := d.nextID
-		d.nextID += 1
-		d.lockOwnerAddrToID[lock.Owner] = id
-		decodedOwnerAddr, err := sdk.AccAddressFromBech32(lock.Owner)
-		if err != nil {
-			return err
-		}
-		d.idToBech32Addr = append(d.idToBech32Addr, lock.Owner)
-		d.idToDecodedAddr = append(d.idToDecodedAddr, decodedOwnerAddr)
-		d.idToDistrCoins = append(d.idToDistrCoins, rewards)
-	}
-	return nil
-}
-
-func (k Keeper) doDistributionSends(ctx sdk.Context, distrs *distributionInfo) error {
-	numIDs := len(distrs.idToDecodedAddr)
-	ctx.Logger().Debug(fmt.Sprintf("Beginning distribution to %d users", numIDs))
-	err := k.bk.SendCoinsFromModuleToManyAccounts(
-		ctx,
-		types.ModuleName,
-		distrs.idToDecodedAddr,
-		distrs.idToDistrCoins)
-	if err != nil {
-		return err
-	}
-	ctx.Logger().Debug("Finished sending, now creating liquidity add events")
-	for id := 0; id < numIDs; id++ {
-		ctx.EventManager().EmitEvents(sdk.Events{
-			sdk.NewEvent(
-				types.TypeEvtDistribution,
-				sdk.NewAttribute(types.AttributeReceiver, distrs.idToBech32Addr[id]),
-				sdk.NewAttribute(types.AttributeAmount, distrs.idToDistrCoins[id].String()),
-			),
-		})
-	}
-	ctx.Logger().Debug(fmt.Sprintf("Finished Distributing to %d users", numIDs))
-	return nil
-}
-
-// distributeInternal runs the distribution logic for a gauge, and adds the sends to
-// the distrInfo computed. It also updates the gauge for the distribution.
-// locks is expected to be the correct set of lock recipients for this gauge.
-func (k Keeper) distributeInternal(
-	ctx sdk.Context, gauge types.Gauge, locks []lockuptypes.PeriodLock, distrInfo *distributionInfo) (sdk.Coins, error) {
-	totalDistrCoins := sdk.NewCoins()
-	lockSum := lockuptypes.SumLocksByDenom(locks, gauge.DistributeTo.Denom)
-
-	if lockSum.IsZero() {
-		return nil, nil
-	}
-
-	remainCoins := gauge.Coins.Sub(gauge.DistributedCoins)
-	remainEpochs := uint64(1)
-	if !gauge.IsPerpetual { // set remain epochs when it's not perpetual gauge
-		remainEpochs = gauge.NumEpochsPaidOver - gauge.FilledEpochs
-	}
-
-	for _, lock := range locks {
-		distrCoins := sdk.Coins{}
-		for _, coin := range remainCoins {
-			// distribution amount = gauge_size * denom_lock_amount / (total_denom_lock_amount * remain_epochs)
-			denomLockAmt := lock.Coins.AmountOfNoDenomValidation(gauge.DistributeTo.Denom)
-			amt := coin.Amount.Mul(denomLockAmt).Quo(lockSum.Mul(sdk.NewInt(int64(remainEpochs))))
-			if amt.IsPositive() {
-				newlyDistributedCoin := sdk.Coin{Denom: coin.Denom, Amount: amt}
-				distrCoins = distrCoins.Add(newlyDistributedCoin)
-			}
-		}
-		distrCoins = distrCoins.Sort()
-		if distrCoins.Empty() {
-			continue
-		}
-		// Update the amount for that address
-		err := distrInfo.addLockRewards(lock, distrCoins)
-		if err != nil {
-			return nil, err
-		}
-
-		totalDistrCoins = totalDistrCoins.Add(distrCoins...)
-	}
-
-	// increase filled epochs after distribution
-	gauge.FilledEpochs += 1
-	gauge.DistributedCoins = gauge.DistributedCoins.Add(totalDistrCoins...)
-	if err := k.setGauge(ctx, &gauge); err != nil {
-		return nil, err
-	}
-
-	return totalDistrCoins, nil
-}
-
-// Distribute coins from gauge according to its conditions
-func (k Keeper) Distribute(ctx sdk.Context, gauges []types.Gauge) (sdk.Coins, error) {
-	distrInfo := newDistributionInfo()
-
-	locksByDenomCache := make(map[string][]lockuptypes.PeriodLock)
-
-	totalDistributedCoins := sdk.Coins{}
-	for _, gauge := range gauges {
-		// All gauges have a precondition of being ByDuration
-		if _, ok := locksByDenomCache[gauge.DistributeTo.Denom]; !ok {
-			locksByDenomCache[gauge.DistributeTo.Denom] = k.getLocksToDistributionWithMaxDuration(
-				ctx, gauge.DistributeTo, time.Millisecond)
-		}
-		// get this from memory instead of hitting iterators / underlying stores.
-		// due to many details of cacheKVStore, iteration will still cause expensive IAVL reads.
-		allLocks := locksByDenomCache[gauge.DistributeTo.Denom]
-		filteredLocks := FilterLocksByMinDuration(allLocks, gauge.DistributeTo.Duration)
-		gaugeDistributedCoins, err := k.distributeInternal(ctx, gauge, filteredLocks, &distrInfo)
-		if err != nil {
-			return nil, err
-		}
-		totalDistributedCoins = totalDistributedCoins.Add(gaugeDistributedCoins...)
-	}
-
-	err := k.doDistributionSends(ctx, &distrInfo)
-	if err != nil {
-		return nil, err
-	}
-
-	k.hooks.AfterEpochDistribution(ctx)
-	return totalDistributedCoins, nil
-}
-
-// GetModuleToDistributeCoins returns sum of to distribute coins for all of the module
-func (k Keeper) GetModuleToDistributeCoins(ctx sdk.Context) sdk.Coins {
-	activeGaugesDistr := k.getToDistributeCoinsFromIterator(ctx, k.ActiveGaugesIterator(ctx))
-	upcomingGaugesDistr := k.getToDistributeCoinsFromIterator(ctx, k.UpcomingGaugesIteratorAfterTime(ctx, ctx.BlockTime()))
-	return activeGaugesDistr.Add(upcomingGaugesDistr...)
-}
-
-// GetModuleDistributedCoins returns sum of distributed coins so far
-func (k Keeper) GetModuleDistributedCoins(ctx sdk.Context) sdk.Coins {
-	activeGaugesDistr := k.getDistributedCoinsFromIterator(ctx, k.ActiveGaugesIterator(ctx))
-	finishedGaugesDistr := k.getDistributedCoinsFromIterator(ctx, k.FinishedGaugesIterator(ctx))
-	return activeGaugesDistr.Add(finishedGaugesDistr...)
 }
 
 // GetGaugeByID Returns gauge from gauge ID
